+++ bluepysnap/circuit_validation.py
@@ -481,7 +481,7 @@ def validate_edge_population(edges_file, name, nodes):
     return []
 
 
-def validate_edges_dict(edges_dict, nodes, skip_slow):
+def validate_edges_dict(edges_dict, nodes, skip_slow, ignore_datatype_errors):
     """Validate an item in the "edges" list.
 
     Args:
@@ -518,7 +518,9 @@ def validate_edges_dict(edges_dict, nodes, skip_slow):
             virtual = False
             if pop_type == "chemical":
                 virtual = _is_source_node_virtual(edges_dict, name, nodes)
-            errors += schemas.validate_edges_schema(edges_file, pop_type, virtual)
+            errors += schemas.validate_edges_schema(
+                edges_file, pop_type, virtual, ignore_datatype_errors
+            )
             if not skip_slow:
                 errors += validate_edge_population(edges_file, name, nodes)
         else:
@@ -527,7 +529,7 @@ def validate_edges_dict(edges_dict, nodes, skip_slow):
     return errors
 
 
-def validate_nodes_dict(nodes_dict, components):
+def validate_nodes_dict(nodes_dict, components, ignore_datatype_errors):
     """Validate an item in the "nodes" list.
 
     Args:
@@ -544,7 +546,9 @@ def validate_nodes_dict(nodes_dict, components):
         nodes_file = nodes_dict["nodes_file"]
 
         if Path(nodes_file).is_file():
-            errors = schemas.validate_nodes_schema(nodes_file, population["type"])
+            errors = schemas.validate_nodes_schema(
+                nodes_file, population["type"], ignore_datatype_errors
+            )
             errors += validate_node_population(nodes_file, population, pop_name)
         else:
             errors.append(BluepySnapValidationError.fatal(f'Invalid "nodes_file": {nodes_file}'))
@@ -552,7 +556,7 @@ def validate_nodes_dict(nodes_dict, components):
     return errors
 
 
-def validate_networks(config, skip_slow):
+def validate_networks(config, skip_slow, ignore_datatype_errors):
     """Validate "networks" part of the config.
 
     Acts as a starting point of validation.
@@ -566,15 +570,17 @@ def validate_networks(config, skip_slow):
 
     for nodes_dict in nodes:
         if "nodes_file" in nodes_dict:
-            errors += validate_nodes_dict(nodes_dict, components)
+            errors += validate_nodes_dict(nodes_dict, components, ignore_datatype_errors)
     for edges_dict in config["networks"].get("edges", []):
         if "edges_file" in edges_dict:
-            errors += validate_edges_dict(edges_dict, nodes, skip_slow)
+            errors += validate_edges_dict(edges_dict, nodes, skip_slow, ignore_datatype_errors)
 
     return errors
 
 
-def validate(config_file, skip_slow, only_errors=False, print_errors=True):
+def validate(
+    config_file, skip_slow, only_errors=False, print_errors=True, ignore_datatype_errors=False
+):
     """Validates Sonata circuit.
 
     Args:
@@ -587,10 +593,10 @@ def validate(config_file, skip_slow, only_errors=False, print_errors=True):
         set: set of errors, empty if no errors
     """
     config = Parser.parse(load_json(config_file), str(Path(config_file).parent))
-    errors = schemas.validate_circuit_schema(config_file, config)
+    errors = schemas.validate_circuit_schema(config_file, config, ignore_datatype_errors)
 
     if "networks" in config:
-        errors += validate_networks(config, skip_slow)
+        errors += validate_networks(config, skip_slow, ignore_datatype_errors)
 
     if _check_partial_circuit_config(config):
         message = (