+++ bluepysnap/edges/edge_population.py
@@ -28,6 +28,7 @@ from more_itertools import first
 from bluepysnap import query, utils
 from bluepysnap.circuit_ids import CircuitEdgeIds, CircuitNodeId
 from bluepysnap.circuit_ids_types import IDS_DTYPE, CircuitEdgeId
+from bluepysnap.edges.edge_population_stats import StatsHelper
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.sonata_constants import DYNAMICS_PREFIX, ConstContainer, Edge
 
@@ -147,6 +148,11 @@ class EdgePopulation:
         """
         return self.get([0], list(self.property_names)).dtypes.sort_index()
 
+    @cached_property
+    def stats(self):
+        """Access edge population stats methods."""
+        return StatsHelper(self)
+
     def container_property_names(self, container):
         """Lists the ConstContainer properties shared with the EdgePopulation.
 
@@ -232,7 +238,7 @@ class EdgePopulation:
         chunk_size = int(1e8)
         for chunk in np.array_split(ids, 1 + len(ids) // chunk_size):
             data = self.get(chunk, properties - unknown_props)
-            res.extend(chunk[query.resolve_ids(data, self.name, queries)])
+            res.extend(chunk[query.resolve_ids(data, self.name, self.type, queries)])
         return np.array(res, dtype=IDS_DTYPE)
 
     def ids(self, group=None, limit=None, sample=None, raise_missing_property=True):
@@ -288,7 +294,7 @@ class EdgePopulation:
             result = result[:limit]
         return utils.ensure_ids(result)
 
-    def get(self, edge_ids, properties):
+    def get(self, edge_ids, properties=None):
         """Edge properties as pandas DataFrame.
 
         Args: