+++ tests/test_spike_report.py
@@ -269,14 +269,14 @@ class TestPopulationSpikeReport:
     @patch(
         test_module.__name__ + ".PopulationSpikeReport.resolve_nodes", return_value=np.asarray([4])
     )
-    def test_get_not_in_report(self, mock):
+    def test_get_not_in_report(self, _):
         pdt.assert_series_equal(self.test_obj.get(4), _create_series([], []))
 
     @patch(
         test_module.__name__ + ".PopulationSpikeReport.resolve_nodes",
         return_value=np.asarray([0, 4]),
     )
-    def test_get_not_in_report(self, mock):
+    def test_get_not_in_report(self, _):
         pdt.assert_series_equal(self.test_obj.get([0, 4]), _create_series([0, 0], [0.2, 1.3]))
 
     def test_node_ids(self):